feat(colors,styles,themes): New public /colors, /styles, and /themes
* Added public `createTheme` function that allows users to generate their own theme based on a mineralTheme, and even supply a different base color.
* Reorganized files according to purpose rather than keeping everything in utils
* Extracted createColorRamp.js and refactored to allow more flexible key names
* Removed styleVariables from public API and merged into createTheme

BREAKING CHANGE: File reorganization - import paths will need updated.  Moved createThemedComponent, mineralTheme, and ThemeProvider from /utils to /themes.  Moved createStyledComponent from /utils to /styles. Removed styleVariables from public API.  Moved color from /utils to /colors.
