+++ src/Card/Card.js
@@ -16,7 +16,7 @@
 
 /* @flow */
 import React from 'react';
-import { createStyledComponent } from '../utils';
+import { createStyledComponent } from '../styles';
 
 type Props = {
   /** Content of the Card. Can be anything, but see [CardBlock](../card-block), [CardImage](../card-image), and [CardTitle](../card-title). */
