+++ src/Card/CardImage.js
@@ -16,7 +16,7 @@
 
 /* @flow */
 import React from 'react';
-import { createStyledComponent } from '../utils';
+import { createStyledComponent } from '../styles';
 import { componentTheme as cardComponentTheme } from './Card';
 
 type Props = Object;
