+++ src/Menu/MenuGroup.js
@@ -16,7 +16,7 @@
 
 /* @flow */
 import React from 'react';
-import { createStyledComponent } from '../utils';
+import { createStyledComponent } from '../styles';
 import MenuGroupTitle from './MenuGroupTitle';
 
 type Props = {
