+++ src/Menu/MenuItem.js
@@ -17,7 +17,7 @@
 /* @flow */
 import React, { cloneElement } from 'react';
 import { withTheme } from 'glamorous';
-import { createStyledComponent, getNormalizedValue, pxToEm } from '../utils';
+import { createStyledComponent, getNormalizedValue, pxToEm } from '../styles';
 import IconDanger from '../Icon/custom/IconDanger';
 import IconSuccess from '../Icon/custom/IconSuccess';
 import IconWarning from '../Icon/custom/IconWarning';
