+++ src/Popover/PopoverTrigger.js
@@ -17,7 +17,7 @@
 /* @flow */
 import React, { Children, cloneElement, Component } from 'react';
 import { Target } from 'react-popper';
-import { createStyledComponent } from '../utils';
+import { createStyledComponent } from '../styles';
 
 type Props = {
   contentId: string,
