+++ src/colors/gray.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  gray_10: '#f5f7fa',
+  gray_20: '#ebeff5',
+  gray_30: '#dde3ed',
+  gray_40: '#c8d1e0',
+  gray_50: '#afbacc',
+  gray_60: '#8e99ab',
+  gray_70: '#707a8a',
+  gray_80: '#58606e',
+  gray_90: '#434a54',
+  gray_100: '#333840'
+};
