+++ src/colors/green.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  green_10: '#e8fcf2',
+  green_20: '#c0fadc',
+  green_30: '#95f5c3',
+  green_40: '#67eba7',
+  green_50: '#40db8d',
+  green_60: '#1fc06f',
+  green_70: '#10a35a',
+  green_80: '#0a8f4c',
+  green_90: '#06783f',
+  green_100: '#046132'
+};
