+++ src/colors/lime.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  lime_10: '#f1fae6',
+  lime_20: '#e7facf',
+  lime_30: '#d2f7a8',
+  lime_40: '#b5f07a',
+  lime_50: '#9be650',
+  lime_60: '#81d42f',
+  lime_70: '#6cbf19',
+  lime_80: '#5aa60d',
+  lime_90: '#4a8c08',
+  lime_100: '#3d7307'
+};
