+++ src/colors/magenta.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  magenta_10: '#f7dae3',
+  magenta_20: '#f7bacd',
+  magenta_30: '#f590b0',
+  magenta_40: '#f06591',
+  magenta_50: '#e63e73',
+  magenta_60: '#d92762',
+  magenta_70: '#c91451',
+  magenta_80: '#b30b43',
+  magenta_90: '#a1083b',
+  magenta_100: '#8d0633'
+};
