+++ src/colors/orange.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  orange_10: '#fce8e1',
+  orange_20: '#fcd0c0',
+  orange_30: '#fab69d',
+  orange_40: '#f59776',
+  orange_50: '#ed774c',
+  orange_60: '#e05b2b',
+  orange_70: '#cf4615',
+  orange_80: '#bd3909',
+  orange_90: '#a83207',
+  orange_100: '#942c06'
+};
