+++ src/colors/purple.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  purple_10: '#eedefa',
+  purple_20: '#e1c5fa',
+  purple_30: '#d5a7fa',
+  purple_40: '#c588f7',
+  purple_50: '#af64ed',
+  purple_60: '#9545d8',
+  purple_70: '#8136bf',
+  purple_80: '#6d2aa3',
+  purple_90: '#5f248f',
+  purple_100: '#511f7a'
+};
