+++ src/colors/sky.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  sky_10: '#e1f3fc',
+  sky_20: '#c0e5fc',
+  sky_30: '#9fd9fc',
+  sky_40: '#79c7f7',
+  sky_50: '#51b3f0',
+  sky_60: '#2f9fe0',
+  sky_70: '#1b8bcc',
+  sky_80: '#0f75b0',
+  sky_90: '#0a6091',
+  sky_100: '#084d75'
+};
