+++ src/colors/teal.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  teal_10: '#e6faf7',
+  teal_20: '#bcf7ee',
+  teal_30: '#8df0de',
+  teal_40: '#5ee6cd',
+  teal_50: '#34d1b7',
+  teal_60: '#20baa3',
+  teal_70: '#15a18e',
+  teal_80: '#0f8c7c',
+  teal_90: '#0c7567',
+  teal_100: '#0a6358'
+};
