+++ src/colors/yellow.js
@@ -15,13 +15,15 @@
  */
 
 /* @flow */
-import React from 'react';
-import { mount } from 'enzyme';
-import ThemeProvider from '../src/themes/ThemeProvider';
-
-export function mountInThemeProvider(Component: React$Element<*>) {
-  const themeProvider = mount(<ThemeProvider>{Component}</ThemeProvider>);
-  const component = themeProvider.find(Component.type);
-
-  return [themeProvider, component];
-}
+export default {
+  yellow_10: '#fcf1dc',
+  yellow_20: '#fce4b6',
+  yellow_30: '#fad48e',
+  yellow_40: '#f7c868',
+  yellow_50: '#f0b241',
+  yellow_60: '#e3a322',
+  yellow_70: '#d19411',
+  yellow_80: '#bd8308',
+  yellow_90: '#a67305',
+  yellow_100: '#916504'
+};
