+++ src/styles/getNormalizedValue.js
