+++ src/styles/pxToEm.js
@@ -15,7 +15,11 @@
  */
 
 /* @flow */
-// Components are built on an 8px grid
-const fontSize_base = 16;
+/**
+  * Helper to convert a px value to ems, relative to the base font size
+  */
+import fontSizeBase from '../themes/fontSizeBase';
 
-export default fontSize_base;
+export default function pxToEm(value: number) {
+  return `${value / fontSizeBase}em`;
+}
